pyproject.toml
[project]
name = "libflowcam"
dynamic = ["version"]
authors = [
  { name="Alex Baldwin", email="contact@alexbaldwin.dev" },
]
license = { text="LGPL-3.0" }
description = "A Python module for interacting with IFCB data."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy >= 2.0.0",
    "pillow >= 8.3.1",
]

[project.urls]
Homepage = "https://github.com/NOC-OI/libflowcam"
Issues = "https://github.com/NOC-OI/libflowcam/issues"

[tool.setuptools.dynamic]

[tool.setuptools]
packages = ["libflowcam"]
package-dir = {"" = "src"}

[tool.setuptools_scm]

[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"
